source/requirements.txt
azure_storage==0.37.0
Flask==2.1.2
numpy==1.22.4
pydub==0.25.1
python_speech_features==0.6
scikit_learn==1.1.1
scipy==1.8.0
torch==1.11.0
ffmpeg
PySoundFile